# Refactor decodeHex in UnescapePure