--- a/benchmarks/Escape.hs
+++ b/benchmarks/Escape.hs
@@ -11,15 +11,21 @@ import qualified "aeson-benchmarks" Data.Aeson.Parser.UnescapeFFI as FFI
 import qualified "aeson-benchmarks" Data.Aeson.Parser.UnescapePure as Pure
 
 import qualified Data.ByteString.Char8 as BS
-
-n :: Int
-n = 10000
-
-input :: BS.ByteString
-input = BS.concat $ replicate n $ BS.pack "\\\""
+import System.Environment (getArgs, withArgs)
 
 main :: IO ()
-main = defaultMain
+main = do
+  args_ <- getArgs
+  let (args, p, n) =
+        case args_ of
+          "--pattern" : p : args_ -> k p args_
+          _ -> k "\\\"" args_
+      k p args_ =
+        case args_ of
+          "--repeat" : n : args_ -> (args_, p, read n)
+          args_ -> (args_, p, 10000)
+      input = BS.concat $ replicate n $ BS.pack p
+  withArgs args $ defaultMain
     [ bench "ffi"  $ whnf FFI.unescapeText input
     , bench "pure" $ whnf Pure.unescapeText input
     ]