Benchmark with a [version of a script](https://github.com/Lysxia/aeson/blob/benchmark.jp.2/slashbench.hs) from #593

Before/after times with a string full of hex encoded characters.

    for X in master my-bench ; ./$X/slashbench pure 10000000 "\\u03a9\\uaBCd\\uEfFa" +RTS -s 2>&1 |tee $X/bench-pure

    ...

      INIT    time    0.000s  (  0.000s elapsed)
      MUT     time    7.633s  (  7.793s elapsed)
      GC      time    1.113s  (  1.281s elapsed)
      EXIT    time    0.000s  (  0.049s elapsed)
      Total   time    8.747s  (  9.123s elapsed)

    ...

      INIT    time    0.000s  (  0.000s elapsed)
      MUT     time    5.510s  (  5.642s elapsed)
      GC      time    1.037s  (  1.261s elapsed)
      EXIT    time    0.000s  (  0.048s elapsed)
      Total   time    6.547s  (  6.951s elapsed)

    ...