@phadej 2017-10-23
Please, add benchmarks into benchmark-suite.