@phadej 2017-10-23
Sorry for brevity, i'm on mobile, check https://github.com/bos/aeson/blob/master/benchmarks/Escape.hs