@Lysxia 2017-10-24
With the default pattern `\"` there is no change, as expected.

With `\uD34d\uBe3F` repeated 2000 times...

Before:

```
benchmarking pure
time                 387.5 μs   (386.4 μs .. 388.6 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 386.3 μs   (385.8 μs .. 386.8 μs)
std dev              1.734 μs   (1.292 μs .. 2.141 μs)
```

After:

```
benchmarking pure
time                 299.8 μs   (297.4 μs .. 303.8 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 298.5 μs   (297.8 μs .. 300.9 μs)
std dev              4.206 μs   (1.045 μs .. 8.720 μs)
```