@bergmark 2017-10-29
Thanks!

This will merge into master which will become 1.3.0.0, which may take a while to release. If anyone would like this sooner please rebase and PR against the 1.2.x branch.