@jprider63 2017-11-18
I'm trying to understand where this speedup is coming from, but am unable to replicate it. I compiled the benchmark suite with `stack bench --stack-yaml stack-bench.yaml`. I then ran `.stack-work-bench/install/x86_64-osx/nightly-2017-10-18/8.2.1/bin/aeson-benchmark-escape --pattern "\\u03a9\\uaBCd\\uEfFa"`:

```
benchmarking pure
time                 2.029 ms   (1.806 ms .. 2.298 ms)
                     0.944 R²   (0.897 R² .. 0.992 R²)
mean                 1.822 ms   (1.779 ms .. 1.929 ms)
std dev              215.4 μs   (104.0 μs .. 408.1 μs)
variance introduced by outliers: 76% (severely inflated)
```

After reverting to the previous version of `decodeHex`, I get the following: 

```
benchmarking pure
time                 2.025 ms   (1.785 ms .. 2.262 ms)
                     0.952 R²   (0.918 R² .. 0.997 R²)
mean                 1.829 ms   (1.792 ms .. 1.929 ms)
std dev              192.9 μs   (111.4 μs .. 364.4 μs)
variance introduced by outliers: 71% (severely inflate
```

I also directly benchmarked both versions of `decodeHex`:

```
benchmarking dataList/match
time                 1.219 ms   (1.217 ms .. 1.220 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.226 ms   (1.223 ms .. 1.229 ms)
std dev              11.79 μs   (7.232 μs .. 20.98 μs)

benchmarking dataList/sub
time                 1.249 ms   (1.244 ms .. 1.255 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.247 ms   (1.244 ms .. 1.252 ms)
std dev              12.99 μs   (10.76 μs .. 16.63 μs)

benchmarking dataList2/match
time                 1.350 ms   (1.263 ms .. 1.411 ms)
                     0.990 R²   (0.987 R² .. 0.999 R²)
mean                 1.263 ms   (1.250 ms .. 1.291 ms)
std dev              60.23 μs   (33.08 μs .. 96.40 μs)
variance introduced by outliers: 36% (moderately inflated)

benchmarking dataList2/sub
time                 1.185 ms   (1.171 ms .. 1.199 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 1.273 ms   (1.248 ms .. 1.299 ms)
std dev              81.98 μs   (69.36 μs .. 103.9 μs)
variance introduced by outliers: 50% (moderately inflated)
```

Any ideas on why I'm not seeing a difference? I'm on a 2015 MBP running OS X 10.13.1.