@Lysxia 2017-11-18
Can I try to run your second benchmark?