@jprider63 2017-11-18
I just pushed the test to this [temporary repository](https://github.com/jprider63/bench-hex). You should just be able to run `make`. They're based on this [blog post](http://tab.snarc.org/posts/haskell/2011-11-15-lookup-tables.html).