@Lysxia 2017-11-18
Interesting, maybe my machine is doing something wrong?

```
stack test --ghc-options -O
benchmarking dataList/match
time                 1.662 ms   (1.575 ms .. 1.748 ms)
                     0.966 R²   (0.936 R² .. 0.986 R²)
mean                 1.728 ms   (1.653 ms .. 1.894 ms)
std dev              373.6 μs   (231.3 μs .. 654.2 μs)
variance introduced by outliers: 93% (severely inflated)

benchmarking dataList/sub
time                 1.795 ms   (1.694 ms .. 1.911 ms)
                     0.974 R²   (0.961 R² .. 0.991 R²)
mean                 1.786 ms   (1.721 ms .. 1.893 ms)
std dev              263.9 μs   (166.1 μs .. 473.5 μs)
variance introduced by outliers: 84% (severely inflated)

benchmarking dataList2/match
time                 2.323 ms   (2.244 ms .. 2.416 ms)
                     0.980 R²   (0.957 R² .. 0.993 R²)
mean                 2.460 ms   (2.389 ms .. 2.602 ms)
std dev              317.8 μs   (208.6 μs .. 565.7 μs)
variance introduced by outliers: 79% (severely inflated)

benchmarking dataList2/sub
time                 1.610 ms   (1.551 ms .. 1.655 ms)
                     0.987 R²   (0.972 R² .. 0.995 R²)
mean                 1.552 ms   (1.520 ms .. 1.622 ms)
std dev              146.9 μs   (80.27 μs .. 266.1 μs)
variance introduced by outliers: 68% (severely inflated)
```