@Lysxia 2017-11-18
No luck with `-O2` either. Here's another machine which agrees with the theory and your results.

```
$ cat /proc/cpuinfo|grep "model name"|head -1
model name	: Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz
```

```
benchmarking dataList/match
time                 1.026 ms   (1.020 ms .. 1.031 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 1.034 ms   (1.028 ms .. 1.046 ms)
std dev              29.36 μs   (12.36 μs .. 57.34 μs)
variance introduced by outliers: 18% (moderately inflated)

benchmarking dataList/sub
time                 1.094 ms   (1.064 ms .. 1.129 ms)
                     0.995 R²   (0.991 R² .. 0.999 R²)
mean                 1.057 ms   (1.044 ms .. 1.079 ms)
std dev              60.16 μs   (44.77 μs .. 91.10 μs)
variance introduced by outliers: 45% (moderately inflated)

benchmarking dataList2/match
time                 1.059 ms   (1.051 ms .. 1.071 ms)
                     0.995 R²   (0.987 R² .. 0.999 R²)
mean                 1.051 ms   (1.034 ms .. 1.098 ms)
std dev              87.71 μs   (50.76 μs .. 139.2 μs)
variance introduced by outliers: 64% (severely inflated)

benchmarking dataList2/sub
time                 991.4 μs   (962.4 μs .. 1.020 ms)
                     0.988 R²   (0.978 R² .. 0.996 R²)
mean                 1.050 ms   (1.010 ms .. 1.112 ms)
std dev              183.8 μs   (116.1 μs .. 259.9 μs)
variance introduced by outliers: 89% (severely inflated)
```